requirements/requirements.txt
# OCR评估框架依赖项

# 核心依赖
PyYAML>=6.0
numpy>=1.20.0

# PaddleOCR支持
paddlepaddle>=2.4.0
paddleocr>=2.6.0
Pillow>=8.0.0

# Qwen多模态模型支持  
lmstudio>=1.0.0

# 数据处理和分析
pandas>=1.3.0
scipy>=1.7.0

# 测试框架
pytest>=6.0.0
pytest-cov>=3.0.0
pytest-mock>=3.0.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
pre-commit>=2.15.0

# 文档生成
Sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0